Refactored to `tui_logic.py` and `project_config.py`
